add player sprite, modify view sizing, tickrate 60, adjust gravity and jump

+++ lib/flappy_elixir/components/y_speed.ex
@@ -3,5 +3,5 @@ defmodule FlappyElixir.Components.YSpeed do
   Documentation for YSpeed components.
   """
   use ECSx.Component,
-    value: :integer
+    value: :float
 end
